Update in Store Sales

+++ API_web_shop/src/main/java/web_shop/api_web_shop/dto/GioHangDTO.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class GioHangDTO {
     private Long id;
     private Long tongSanPham;
-    private Long tongSoTien; // tong so tien cua tat ca cac gio hang chi tiet trong gio hang
+    private Double tongSoTien; // tong so tien cua tat ca cac gio hang chi tiet trong gio hang
 
     private TaiKhoanDTO taiKhoan;
